+++ lib/widgets/hoverable_widget.dart
@@ -6,7 +6,7 @@ class HoverableWidget extends StatefulWidget {
   final BorderRadius? borderRadius;
   final EdgeInsetsGeometry? padding;
   final Function()? onTapDown;
-  final Function()? onDoubleTap;
+  // final Function()? onDoubleTap;
   final Function()? onTap;
 
   const HoverableWidget({
@@ -16,7 +16,7 @@ class HoverableWidget extends StatefulWidget {
     this.borderRadius,
     this.padding,
     this.onTapDown,
-    this.onDoubleTap,
+    // this.onDoubleTap,
     this.onTap,
   });
 
@@ -47,9 +47,9 @@ class _HoverableWidgetState extends State<HoverableWidget> {
         onTap: () {
           if (widget.onTap != null) widget.onTap!();
         },
-        onDoubleTap: () {
-          if (widget.onDoubleTap != null) widget.onDoubleTap!();
-        },
+        // onDoubleTap: () {
+        //   if (widget.onDoubleTap != null) widget.onDoubleTap!();
+        // },
         onTapDown: (event) {
           if (widget.onTapDown != null) widget.onTapDown!();
         },